refactor(grocery): rename GroceryFragment to MyKitchen.
