+++ app/src/main/res/layout/fragment_grocery_list.xml
@@ -7,5 +7,5 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layoutManager="LinearLayoutManager"
-    tools:context=".layouts.mykitchen.GroceryFragment"
+    tools:context=".layouts.mykitchen.MyKitchen"
     tools:listitem="@layout/fragment_grocery" />
