--- app/src/main/res/layout/fragment_my_kitchen.xml
@@ -1,58 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/kitchen_frame"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".layouts.mykitchen.MyKitchen">
-
-
-    <android.support.v7.widget.CardView
-        android:id="@+id/card"
-        android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:layout_marginLeft="@dimen/cardview_margin"
-        android:layout_marginTop="@dimen/cardview_margin"
-        android:layout_marginRight="@dimen/cardview_margin"
-        app:cardCornerRadius="@dimen/cardview_radius">
-
-        <ImageView
-            android:id="@+id/card_view_image"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:contentDescription="@string/app_name"
-            android:scaleType="centerCrop"
-            android:visibility="visible"
-            app:srcCompat="@drawable/card_my_groceries_low" />
-
-        <TextView
-            android:id="@+id/card_text"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:layout_marginRight="@dimen/cardview_margin"
-            android:layout_marginEnd="@dimen/cardview_margin"
-            android:layout_marginBottom="@dimen/cardview_margin"
-            android:gravity="bottom|end"
-            android:text="@string/title_groceries"
-            android:textColor="@color/colorWhite"
-            android:textSize="@dimen/title_size"
-            android:textStyle="bold" />
-
-    </android.support.v7.widget.CardView>
-
-    <RelativeLayout
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:gravity="bottom">
-
-        <Button
-            style="@style/Widget.AppCompat.Button.Colored"
-            android:id="@+id/generate_recipe_button"
-            android:layout_width="fill_parent"
-            android:layout_height="wrap_content"
-            android:text="@string/kitchen_generate_recipe" />
-
-    </RelativeLayout>
-
-</FrameLayout>
